+++ CMakeLists.txt
@@ -23,7 +23,9 @@ add_library(co_lib
         include/co/impl/waiting_queue.cpp
         include/co/thread.cpp
         include/co/this_thread.cpp
-        include/co/mutex.cpp)
+        include/co/mutex.cpp
+        include/co/net/tcp_listener.cpp
+        include/co/net/impl/uv_tcp_ptr.cpp)
 
 target_link_libraries(co_lib PUBLIC ${CONAN_LIBS})
 target_link_directories(co_lib PUBLIC ${CONAN_LIB_DIRS})
