+++ include/co/impl/uv_handler.hpp
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <cassert>
 #include <memory>
 #include <uv.h>
